refactor: move chat avatar
